Update the likes module
