+++ src/module/card.js
@@ -1,6 +1,6 @@
 import modal from './modal.js';
 import { likeurl } from './url.js';
-import requestLikes from './lrequesLike.js';
+import { requestLikes } from './getData.js';
 
 const card = (movies, data) => {
   const card = document.querySelector('.card-container');
